Refactor #247 - move `colorPressButton` to `androidx.component-ext`
