+++ app/src/main/java/com/nlab/reminder/domain/feature/home/ui/BottomContainerButton.kt
@@ -32,7 +32,7 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.nlab.reminder.R
-import com.nlab.reminder.core.android.designsystem.component.ColorPressButton
+import com.nlab.reminder.core.androidx.compose.ui.component.ColorPressButton
 import com.nlab.reminder.core.androidx.compose.ui.throttle
 import com.nlab.reminder.core.designsystem.compose.theme.ReminderTheme
